review: move extractIP() to _linux file
