--- pkg/util/system/network_stub.go
@@ -1,13 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build !linux
-// +build !linux
-
-package system
-
-func ParseProcessIPs(procPath string, pid int, filterFunc func(string) bool) ([]string, error) {
-	return nil, nil
-}
