+++ pkg/workloadmeta/collectors/internal/containerd/container_builder.go
@@ -9,17 +9,14 @@
 package containerd
 
 import (
-	"errors"
 	"fmt"
 	"time"
 
 	"github.com/containerd/containerd"
 	"github.com/containerd/containerd/errdefs"
 
-	"github.com/DataDog/datadog-agent/pkg/config"
 	cutil "github.com/DataDog/datadog-agent/pkg/util/containerd"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	"github.com/DataDog/datadog-agent/pkg/util/system"
 	"github.com/DataDog/datadog-agent/pkg/workloadmeta"
 )
 
@@ -110,39 +107,4 @@ func extractStatus(status containerd.ProcessStatus) workloadmeta.ContainerStatus
 	}
 
 	return workloadmeta.ContainerStatusUnknown
-}
-
-// extractIP gets the IP of a container.
-//
-// The containerd client does not expose the IPs, that's why we use the helpers
-// that we have in the "system" package to extract that information from
-// "/proc".
-//
-// A current limitation is that if a container exposes multiple IPs, this
-// function just returns one of them. That means that if a container is attached
-// to multiple networks this might not work as expected.
-func extractIP(container containerd.Container, containerdClient cutil.ContainerdItf) (string, error) {
-	taskPids, err := containerdClient.TaskPids(container)
-	if err != nil {
-		return "", err
-	}
-
-	if len(taskPids) == 0 {
-		return "", errors.New("no PIDs found")
-	}
-
-	IPs, err := system.ParseProcessIPs(
-		config.Datadog.GetString("container_proc_root"),
-		int(taskPids[0].Pid), // Any PID of the container should work
-		func(ip string) bool { return ip != "127.0.0.1" },
-	)
-	if err != nil {
-		return "", err
-	}
-
-	if len(IPs) == 0 {
-		return "", errors.New("no IPs found")
-	}
-
-	return IPs[0], nil
 }
